+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `properties.autoKeyword` metric
 - Added `mediaQueries.features.deprecated` metric
 - Added `mediaQueries.types.deprecated` metric
+- Added `properties.uniqueRatio` metric
+- Added `properties.shorthandsRatio` metric
 
 ### Changed
 - Reorganized `properties.explicitDefaultingKeywords` to: